version 0.0.4 - finding the location of the pdf2txt before using it
